Fixed MAtrix dataset. Changed Dictionary<int,float> to UserPreference class

+++ BrianDennis.INFDTA01.Opdracht1/Models/UserPreference.cs
@@ -3,8 +3,8 @@
     public class UserPreference 
     {
         public int MovieId { get; set; }
-        public string Description { get; set; }
-        public int  Rating { get; set; }
+        public string Description { get; set; } 
+        public float  Rating { get; set; }
         public string Genre { get; set; }
     }
 }
